map_backend/requirements.txt
# =============================================================================
# MAP Backend — 完整依赖清单（由代码扫描生成 + 显式底层驱动）
# Python 3.10+  |  pip install -r requirements.txt
#
# 【重要】依赖必须装进「运行 uvicorn 的同一个解释器」里（推荐项目 venv）：
#   cd backend
#   .\venv\Scripts\pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple
# 若只用全局/Anaconda 的 pip 安装，而 uvicorn 用 backend\venv，会出现 ModuleNotFoundError。
#
# 扫描范围：backend/**/*.py
# 第三方包映射：
#   fastapi, starlette          → Web 框架
#   uvicorn[standard]           → ASGI 服务器（含 uvloop / httptools / watchfiles）
#   sqlalchemy                  → ORM + 异步引擎（隐式依赖 greenlet）
#   greenlet                    → SQLAlchemy asyncio 必需（显式列出避免环境残缺）
#   aiomysql                    → mysql+aiomysql 异步驱动
#   pymysql                     → PyMySQL：aiomysql 底层同步协议栈 / 同步工具脚本
#   pydantic, pydantic-settings → 校验与配置
#   python-jose[cryptography]   → JWT（jose 包名）
#   cryptography                → JWT 非对称算法 / TLS（python-jose 强烈建议）
#   passlib[bcrypt], bcrypt     → 密码哈希
#   httpx, httpcore, certifi    → 异步 HTTP（httpcore/certifi 为 httpx 链，显式防残缺）
#   tenacity                    → Adapter 重试
#   redis[hiredis], hiredis    → 异步 Redis + C 解析加速
#   alembic                     → 数据库迁移（与 SQLAlchemy 配套）
#   pytest, pytest-asyncio      → 测试
# =============================================================================

# --- Web / ASGI ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
starlette>=0.37.0

# --- ORM / 数据库 ---
sqlalchemy[asyncio]>=2.0.36
greenlet>=3.0.0
alembic>=1.13.0

# --- MySQL / TDSQL 异步栈（显式三层，避免 ModuleNotFoundError）---
aiomysql>=0.2.0
pymysql>=1.1.0

# --- 本地 SQLite（MAP_USE_SQLITE=true 时由 AsyncEngine 使用）---
aiosqlite>=0.20.0

# --- 配置 / 校验 ---
pydantic>=2.7.0
pydantic-settings>=2.3.0

# --- 安全（JWT + 密码）---
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# --- HTTP 客户端 / 重试 ---
httpx>=0.27.0
httpcore>=1.0.0
certifi>=2024.0.0
tenacity>=8.3.0

# --- Redis ---
redis[hiredis]>=5.0.0
hiredis>=2.3.0

# --- 测试 ---
pytest>=8.0.0
pytest-asyncio>=0.23.0

# --- 可选：启用 RocketMQ 生产者时取消注释 ---
# rocketmq-client-python>=2.0.0

# --- 可选：若将来使用异步 PostgreSQL，取消注释 ---
# asyncpg>=0.29.0
